+++ 03_Core_Cpp/headers/bufferHeaders.hpp
@@ -2,13 +2,13 @@
 #define BUFFER_HEADER_H
 
 struct ControllerBufferHeader {
-    double systemStatus;
-    double socketSymbolCount;
-    double tbtSocketSymbolCount;
-    double apiSymbolCount;
-    double marketDepthCount;
-    double signal;
-    double action;
+    int systemStatus;
+    int socketSymbolCount;
+    int tbtSocketSymbolCount;
+    int apiSymbolCount;
+    int marketDepthCount;
+    int signal;
+    int action;
 };
 
 struct OptionChainBufferHeader {
